Commands -> Module - Renamed commands folder to module - Makes sense when loading and unloading - Renamed cog_loader to loader   - Added some error handling   - Cannot unload yourself. Sorry bucko, no sudoku
